Spring Break 2024
If a DVD checks out Mar 4-17, then it is due Mar 27 (as if it checked out Mar 18).

Else If a DVD is due Mar 8-17, then it is due Mar 18.

If a book is checked out Mar 4-17, then it is due Apr 8 (as if it were checked out Mar 18).

Else if a book is due Mar 8-17, then it is due Mar 18.