Revert some closed days to open
Because the MNPS closed date is about whether a student can turn in a book on that day, not whether LL circ staff can/should check something out that day, I have re-opened the day of the final due date, 5/11/2026, and the two days preceding. 5/5/2025 remains closed; it's a stockpiled PD day and the students do not report to school.